+++ crates/stackable-operator/CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to this project will be documented in this file.
 
 - Bump GitHub workflow actions ([#772]).
 - Revert `zeroize` version bump ([#772]).
+- Support specifying externalTrafficPolicy in Services created by listener-operator ([#562]).
 
+[#562]: https://github.com/stackabletech/operator-rs/pull/562
 [#772]: https://github.com/stackabletech/operator-rs/pull/772
 
 ## [0.67.0] - 2024-04-25